// frontend/contracts/core/v1/models/NotificationModel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type NotificationModel = {
    id: string;
    user_id: string;
    type: string;
    title: string;
    body: string;
    link?: string;
    data?: Record<string, any>;
    is_read: boolean;
    created_at: string;
};